Implement text entry key API for accessibility services in AOSP Keyboard
From Android Q, Talkback supports lift-to-type feature if the node claims
it is a text entry key via setTextEntryKey(). We implement this API to
show how this API is applied.

This CL uses AccessibilityNodeInfoCompat instead of AccessibilityNodeInfo
so that the same functionality can be used even on pre-Q devices when
the AccessibilityService supports it.
With that, this CL removes the legacy code of lift-to-type feature,
which was implemented in the AOSP Keyboard side.

Bug: 131644969

Test: manual - enable Talkback suporting lift-to-type,check Talback
perform click action when finger is lifted.

Change-Id: I1ec2928f5a9ba0bde999b09d4c0b9c922f179a2a