Revert "LatinIME: Do we really still want Holo?"
This reverts commit 13764c70334d157332d73a21a0af3ba101ed7034.

Reason for revert: There is no reason to remove Holo just because it is 'old'. A lot of people still like Holo. LineageOS has a lot of customization features, so removing a keyboard-theme for no reason is not a good idea. Since Holo was already in 17.1-branch, it should be there at least until end of 17.1 as long as there is no technical reason.

Change-Id: Ie57f75e23851f780ce302723d95d9f6b2feaf6c1
Signed-off-by: Pranav Vashi <neobuddy89@gmail.com>
Signed-off-by: SirRGB <sirrgb@proton.me>